+++ test/ReaderEither.ts
@@ -188,13 +188,6 @@ describe('ReaderEither', () => {
     assert.deepStrictEqual(_.asks((r: { readonly a: number }) => r.a)({ a: 1 }), E.right(1))
   })
 
-  it('local', () => {
-    assert.deepStrictEqual(
-      _.local((n: number) => ({ a: n }))((r: { readonly a: number }) => E.right(r.a))(1),
-      E.right(1)
-    )
-  })
-
   it('getApplicativeReaderValidation', () => {
     const A = _.getApplicativeReaderValidation(monoidString)
     assert.deepStrictEqual(Apply.sequenceT(A)(_.left('a'), _.left('b'))(null), E.left('ab'))